+++ test/test_helpers/pact_helper.rb
@@ -2,6 +2,7 @@ PUBLISHING_API_PORT = 3001
 ORGANISATION_API_PORT = 3002
 BANK_HOLIDAYS_API_PORT = 3003
 ACCOUNT_API_PORT = 3004
+LINK_CHECKER_API_PORT = 3005
 
 def publishing_api_host
   "http://localhost:#{PUBLISHING_API_PORT}"
@@ -17,6 +18,10 @@ end
 
 def account_api_host
   "http://localhost:#{ACCOUNT_API_PORT}"
+end
+
+def link_checker_api_host
+  "http://localhost:#{LINK_CHECKER_API_PORT}"
 end
 
 Pact.service_consumer "GDS API Adapters" do
@@ -41,6 +46,12 @@ Pact.service_consumer "GDS API Adapters" do
   has_pact_with "Account API" do
     mock_service :account_api do
       port ACCOUNT_API_PORT
+    end
+  end
+
+  has_pact_with "Link Checker API" do
+    mock_service :link_checker_api do
+      port LINK_CHECKER_API_PORT
     end
   end
 end